chore: move sync-integrations to config package
